# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ExamSummary {
  id: string;
  title: string;
  subject: string;
  gradeLevel: string;
  description: string;
  questionCount: number;
  durationMinutes: number | null;
  difficulty: string;
  types: string[];
  status: string;
}

export interface ExamSection {
  title: string;
  description: string;
}

export interface Question {
  id: string;
  number: number;
  type: string;
  text: string;
  points: number;
  section: string;
  concepts: string[];
  options: string[];
  imageUrl?: string | null;
}

export type ExamDetail = ExamSummary & {
  sections: ExamSection[];
  questions: Question[];
};

export interface ResultSummary {
  id: string;
  examId: string;
  examTitle: string;
  subject: string;
  grade: string;
  scorePercent: number;
  completedAt: string;
}

export interface Dashboard {
  activeExams: ExamSummary[];
  recentResults: ResultSummary[];
}

export interface AdminSession {
  authenticated: boolean;
  username: string | null;
  csrfToken: string | null;
}

export interface AdminLoginInput {
  /** @minLength 1 */
  username: string;
  /** @minLength 1 */
  password: string;
}

export interface AttemptStartInput {
  examId: string;
}

export interface AttemptSaveInput {
  questionId: string;
  answer: string;
  marked?: boolean;
}

export interface SavedAnswer {
  questionId: string;
  answer: string;
  marked: boolean;
}

export interface Attempt {
  id: string;
  examId: string;
  startedAt: string;
  status: string;
  currentQuestion: number;
  answers: SavedAnswer[];
}

export interface QuestionResult {
  questionId: string;
  status: string;
  feedback: string;
}

export interface DimensionScore {
  label: string;
  score: number;
}

export type ExamResult = ResultSummary & {
  strengths: string[];
  practiceAreas: string[];
  goldenFeedback: string;
  nextSteps: string[];
  questionResults: QuestionResult[];
  dimensionScores: DimensionScore[];
};

export interface LearningProfile {
  subject: string;
  dimensions: DimensionScore[];
  focusLabel: string;
  focusText: string;
}

export interface AdminSummary {
  totalExams: number;
  totalQuestions: number;
  totalAttempts: number;
  averageScore: number;
  mostCompletedExam: string;
  aiUsage: number;
}

export type AdminExam = ExamSummary & {
  createdAt: string;
  updatedAt: string;
};

export interface ExamUpdate {
  title?: string;
  description?: string;
  difficulty?: string;
  durationMinutes?: number | null;
  status?: string;
}