# lib/api-zod/src/generated/types/adminSession.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminSession {
  authenticated: boolean;
  username: string | null;
  csrfToken: string | null;
}